Create Proper_Noun_Extractor.py
A quick script to obtain all the capitalized words in SAIS Captions.txt. grouped together if they appeared consecutively in the file, and otherwise delimited by tabs.
The output is relatively dirty, but as the words will have to be looked over by hand to classify them, it should be relatively easy for the reader to determine whether two phrases that may not have been tab-delimited are meant to be separated.